+++ app/src/androidTest/java/com/example/drinktracker/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.drinktracker
+package com.yambo.drinktracker
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.drinktracker", appContext.packageName)
+        assertEquals("com.yambo.drinktracker", appContext.packageName)
     }
 }
