--- app/src/main/java/com/example/drinktracker/models/BottleSize.kt
@@ -1,10 +1,0 @@
-package com.example.drinktracker.models
-
-data class BottleSize(
-    val size: Float,
-    val unit: Unit
-) {
-    override fun toString(): String {
-        return "$size ${unit.unitString}"
-    }
-}
