--- app/src/main/java/com/example/drinktracker/models/Unit.kt
@@ -1,9 +1,0 @@
-package com.example.drinktracker.models
-
-enum class Unit(val unitString: String) {
-    OUNCE("ounce"),
-    LITER("liter"),
-    GALLON("gallon"),
-    MILLILITER("milliliter"),
-    FLUID_OUNCE("fluid ounce")
-}
