--- app/src/main/java/com/example/drinktracker/models/Water.kt
@@ -1,13 +1,0 @@
-package com.example.drinktracker.models
-
-import com.google.gson.Gson
-
-data class Water(
-    val companyName: String,
-    val bottleSize: BottleSize
-) {
-    override fun toString(): String {
-        val gson = Gson()
-        return gson.toJson(this)
-    }
-}
