--- app/src/main/java/com/example/drinktracker/repository/definitions/WaterCompanyRepositoryInterface.kt
@@ -1,9 +1,0 @@
-package com.example.drinktracker.repository.definitions
-
-import com.example.drinktracker.models.BottleSize
-
-interface WaterCompanyRepositoryInterface {
-    fun getAllCompanyNames() : Array<String>
-    fun getBottleSizesForCompany(companyName: String) : Array<BottleSize>
-    fun getAllBottleSizes() : Array<BottleSize>
-}
