--- app/src/main/java/com/example/drinktracker/repository/definitions/WaterConsumptionRepositoryInterface.kt
@@ -1,8 +1,0 @@
-package com.example.drinktracker.repository.definitions
-
-import com.example.drinktracker.models.Water
-
-interface WaterConsumptionRepositoryInterface {
-    fun getAllWaterForUser(id: String): Array<Water>
-    fun addWaterToUser(id: String, water: Water)
-}
