--- app/src/main/java/com/example/drinktracker/utils/Constants.kt
@@ -1,6 +1,0 @@
-package com.example.drinktracker.utils
-
-object Constants {
-    const val LOCAL_STORAGE_WATER_DIRECTORY_NAME = "local_storage_water_directory_name"
-    const val LOCAL_STORAGE_WATER_FILE_NAME = "local_storage_water_file_name"
-}
