+++ app/src/main/java/com/yambo/drinktracker/models/BottleSize.kt
@@ -1,0 +1,10 @@
+package com.yambo.drinktracker.models
+
+data class BottleSize(
+    val size: Float,
+    val unit: Unit
+) {
+    override fun toString(): String {
+        return "$size ${unit.unitString}"
+    }
+}
