+++ app/src/main/java/com/yambo/drinktracker/models/Unit.kt
@@ -1,0 +1,9 @@
+package com.yambo.drinktracker.models
+
+enum class Unit(val unitString: String) {
+    OUNCE("ounce"),
+    LITER("liter"),
+    GALLON("gallon"),
+    MILLILITER("milliliter"),
+    FLUID_OUNCE("fluid ounce")
+}
