+++ app/src/main/java/com/yambo/drinktracker/models/Water.kt
@@ -1,0 +1,13 @@
+package com.yambo.drinktracker.models
+
+import com.google.gson.Gson
+
+data class Water(
+    val companyName: String,
+    val bottleSize: BottleSize
+) {
+    override fun toString(): String {
+        val gson = Gson()
+        return gson.toJson(this)
+    }
+}
