+++ app/src/main/java/com/yambo/drinktracker/repository/definitions/WaterConsumptionRepositoryInterface.kt
@@ -1,0 +1,8 @@
+package com.yambo.drinktracker.repository.definitions
+
+import com.yambo.drinktracker.models.Water
+
+interface WaterConsumptionRepositoryInterface {
+    fun getAllWaterForUser(id: String): Array<Water>
+    fun addWaterToUser(id: String, water: Water)
+}
