+++ app/src/main/res/layout/activity_drink_tracker.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/main_fragment_container_view"
-        android:name="com.example.drinktracker.views.ConsumedWaterListFragment"
+        android:name="com.yambo.drinktracker.views.ConsumedWaterListFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
